requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
asyncpg>=0.29.0
pgvector>=0.2.4

# ML Models
sentence-transformers>=2.3.0
llama-cpp-python>=0.2.0
transformers>=4.36.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
